# mempool: Keep cache hashmap and linked list in sync