--- a/CHANGELOG_PENDING.md
+++ b/CHANGELOG_PENDING.md
@@ -15,3 +15,5 @@ FEATURES:
 IMPROVEMENTS:
 
 BUG FIXES:
+- [mempool] No longer possible to fill up linked list without getting caching 
+benefits [#2180](https://github.com/tendermint/tendermint/issues/2180)