--- a/mempool/mempool.go
+++ b/mempool/mempool.go
@@ -2,6 +2,7 @@ package mempool
 
 import (
 	"bytes"
+	"container/list"
 	"fmt"
 	"sync"
 	"sync/atomic"
@@ -487,8 +488,8 @@ type txCache interface {
 type mapTxCache struct {
 	mtx  sync.Mutex
 	size int
-	map_ map[string]*clist.CElement
-	list *clist.CList // to remove oldest tx when cache gets too big
+	map_ map[string]*list.Element
+	list *list.List // to remove oldest tx when cache gets too big
 }
 
 var _ txCache = (*mapTxCache)(nil)
@@ -497,15 +498,15 @@ var _ txCache = (*mapTxCache)(nil)
 func newMapTxCache(cacheSize int) *mapTxCache {
 	return &mapTxCache{
 		size: cacheSize,
-		map_: make(map[string]*clist.CElement, cacheSize),
-		list: clist.New(),
+		map_: make(map[string]*list.Element, cacheSize),
+		list: list.New(),
 	}
 }
 
 // Reset resets the cache to an empty state.
 func (cache *mapTxCache) Reset() {
 	cache.mtx.Lock()
-	cache.map_ = make(map[string]*clist.CElement, cache.size)
+	cache.map_ = make(map[string]*list.Element, cache.size)
 	cache.list.Init()
 	cache.mtx.Unlock()
 }
@@ -524,9 +525,9 @@ func (cache *mapTxCache) Push(tx types.Tx) bool {
 		popped := cache.list.Front()
 		poppedTx := popped.Value.(types.Tx)
 		delete(cache.map_, string(poppedTx))
-		cache.list.Remove(popped)
-		popped.DetachPrev()
-		popped.DetachNext()
+		if popped != nil {
+			cache.list.Remove(popped)
+		}
 	}
 	cache.list.PushBack(tx)
 	cache.map_[string(tx)] = cache.list.Back()
@@ -539,9 +540,9 @@ func (cache *mapTxCache) Remove(tx types.Tx) {
 	stx := string(tx)
 	popped := cache.map_[stx]
 	delete(cache.map_, stx)
-	cache.list.Remove(popped)
-	popped.DetachPrev()
-	popped.DetachNext()
+	if popped != nil {
+		cache.list.Remove(popped)
+	}
 
 	cache.mtx.Unlock()
 }