@ValarDragon 2018-08-09
The benefit of this change is space savings + fixing bugs with cache eviction. Computation time will increase due to usage of the clist, so making a benchmark is a good idea. 

Is there a good way to benchmark size? Just println unsafe size of?