@melekes 2018-08-09
`go test -memprofile`?