@ValarDragon 2018-08-09
before:
```
BenchmarkCacheInsertTime-8   	 5000000	       385 ns/op	      88 B/op	       3 allocs/op
BenchmarkCacheRemoveTime-8   	10000000	       152 ns/op	       0 B/op	       0 allocs/op
```
after:
```
BenchmarkCacheInsertTime-8   	 1000000	      1027 ns/op	     360 B/op	       7 allocs/op
BenchmarkCacheRemoveTime-8   	 2000000	       860 ns/op	     112 B/op	       2 allocs/op
```
So it looks like we're getting a performance hit due to the concurrent list. However this performance hit is neglible compared to the signature verification of saving a checkTx, and we are now saving memory from elements we remove from the cache. (The before case didn't remove it at all from the linked list. If we used the O(N) method to do that properly, it would certainly be slower)