@melekes 2018-08-09
> However this performance hit is neglible compared to the signature verification of saving a checkTx, and we are now saving memory from elements we remove from the cache.

I tend to agree. Plus, if there are usually no "bad" txs in the network (e.g. private deployment), cache can be always disabled.