@ValarDragon 2018-08-11
Now that we're back to normal lists, here are the new benchmarks:
```
BenchmarkCacheInsertTime-8   	 5000000	       398 ns/op	      88 B/op	       3 allocs/op
BenchmarkCacheRemoveTime-8   	10000000	       179 ns/op	       0 B/op	       0 allocs/op
```
basically just adds 20ns to removal time.